Organized the matches package
