--- src/main/java/team/gif/gearscout/matches/MatchPreprocessor.java
@@ -1,7 +1,0 @@
-package team.gif.gearscout.matches;
-
-public interface MatchPreprocessor {
-
-	void process(NewMatch match);
-
-}
