+++ src/main/java/team/gif/gearscout/matches/preprocessor/MatchPreprocessor.java
@@ -1,0 +1,9 @@
+package team.gif.gearscout.matches.preprocessor;
+
+import team.gif.gearscout.matches.model.NewMatch;
+
+public interface MatchPreprocessor {
+
+	void process(NewMatch match);
+
+}
